Add a 'Let's Check' Button to improve UX
If users use file selector to choose some files, they now need to press
the button to confirm and start processing.  This seems to have a better
UX.

Also let's show the file names we processed in the results area, so user
knows their files are processed (to be more responsive).